requirements.txt
genanki==0.13.0
pdfminer.six==20221105
pytesseract==0.3.10
Pillow==10.0.0
python-docx==0.8.11
pdf2image==1.16.3  # For better PDF handling
openai==0.25.0
python-dotenv==1.0.0  # For environment variables
